external/my_stories.csv: Fix formatting in some stories
Some stories did not always follow the convention of (*{X}*) to refer to keyword X. That caused the koohii-stories page to poorly format such stories. Fix some of these and then add a story for frame #2584 (pheasant) which is now properly referred to and therefore must appear in the set of exported stories.