Switch the onomatopoeia page to being generated
Now that the python script that generates the onomatopoeia page is "good enough", switch to using it to build the page.

study-material/study.make: New makefile that invokes the python script that builds the generated page.

study-material/Makefile: Invoke the new makefile as a component.

study-material/progress/Makefile: Remove the build of the previous "static" version of this page.

study-material/progress/study-material.jhtml: Link to the generated page rather than the former static one.