scripts/jpdb_review_all_vocab.py: Script to perform all outstanding reviews
JPDB.io has a limit of 10,000 words in "never forget", at least for now.

So the only way left to know what is genuinely new vocab is to review everything that is outstanding and then add more text and see which words are new. Then pick up those words and finally review them.

This script peforms the review and marks words as Easy.