scripts/jpdb_review_all_vocab.py: Add code to do initial review of New vocab
I've got all the Due vocab out of the way, so now add a function to handle New vocab.
At some point (probably by tomorrow!) I'll have to make the code bright enough to run through all the Due vocab and then run through all the New vocab, but that's a battle for the future.