scripts/koohii-stories.py: Update variable name for one of the kanji maps
The name 'frame_number_map' must have meant something to me when I chose it but coming back to this code I find that it means less now than it did then.
Every entry repreesents a specific kanji and each entry contains a list of all the kanji that reference this kanji.
As an example the kanji 日 is used in the stories for 冒, 明, 唱 and 早 as well as many others, so eahc of these should be in the list indexed by 日.
I've renamed this kanji_referenced_by_map. It's not perfect but I'm soon to add a map of the kanji that a given kanji refers to, which is the inverse of this one, and I need to keep them straight. 'frame_number_map' was not helping!