scripts/koohii-stories.py: Fill in column 5 of Kanji Stories
Column 6 is "other kanji that use this kanji as one of their building blocks (i.e. in their stories)".
Column 5 is "other kanji used by this kanji as one of its building blocks (i.e. the other kanji appear in this kanji's story)."

Column 6 was already filled in.
Column 5 is now filled in.

So clicking on a kanji in column 6 takes you to a kanji story that includes the kanji in this row.
Clicking on a kanji in column 5 takes you to a story for a kanji that this kanji uses as an aide memoire.